+++ src/main/java/com/jhlabs/image/ConstantGradientBlend.java
@@ -1,10 +1,8 @@
 package com.jhlabs.image;
 
-import java.awt.Color;
-
 public class ConstantGradientBlend implements GradientBlend{
 	@Override
-	public int GradientBlendOperation(int t){
+	public float GradientBlendOperation(float t){
 		return 0;
 	}
 }
