+++ src/main/java/com/jhlabs/image/SplineGradientBlend.java
@@ -1,10 +1,8 @@
 package com.jhlabs.image;
 
-import java.awt.Color;
-
 public class SplineGradientBlend implements GradientBlend{
 	@Override
-	public int GradientBlendOperation(int t){
+	public float GradientBlendOperation(float t){
 		return ImageMath.smoothStep(0.15f, 0.85f, t);
 	}
 }
